# docs/context/api/openapi.yaml
openapi: 3.1.0
info:
  title: UA Workflow Platform API
  version: 1.0.0-t033
  description: |
    Pure-v1 authoritative API baseline frozen by T-033.

    Contract rules:
    - `POST /v1/agents/{agentId}/runs` is the only production ingress.
    - `POST /v1/runs` is studio/debug-only direct version start.
    - Approval and interaction resumes are split by authoritative identities.
    - Governance-change request endpoints remain auxiliary proposal storage and are not authoritative runtime approval state.
servers: []
...
paths:
...
  /v1/workflows/{workflowId}:
    get:
      tags: [Workflow Templates]
      summary: Get a workflow template and its published versions.
      parameters:
        - $ref: "#/components/parameters/workflowId"
      responses:
        "200":
          description: Workflow detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetailResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/workflow-drafts:
    get:
      tags: [Workflow Drafts]
      summary: List workflow drafts visible to the current control-plane session.
      responses:
        "200":
          description: Draft list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftListResponse"
    post:
      tags: [Workflow Drafts]
      summary: Create a workflow draft.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftCreateRequest"
      responses:
        "200":
          description: Draft created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftMutateResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/workflow-drafts/{draftId}:
    get:
      tags: [Workflow Drafts]
      summary: Get a workflow draft.
      parameters:
        - $ref: "#/components/parameters/draftId"
      responses:
        "200":
          description: Draft detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftDetailResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/workflow-drafts/{draftId}/revisions:
    get:
      tags: [Workflow Drafts]
      summary: List revisions for a workflow draft.
      parameters:
        - $ref: "#/components/parameters/draftId"
      responses:
        "200":
          description: Draft revisions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftRevisionListResponse"
  /v1/workflow-drafts/{draftId}/focus:
    post:
      tags: [Workflow Drafts]
      summary: Focus a draft in the current authoring session.
      parameters:
        - $ref: "#/components/parameters/draftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftFocusRequest"
      responses:
        "200":
          description: Draft focused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftFocusResponse"
  /v1/workflow-drafts/{draftId}/intake:
    post:
      tags: [Workflow Drafts]
      summary: Append authoring intake to a draft.
      description: Authoring helper only. Not production ingress.
      parameters:
        - $ref: "#/components/parameters/draftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftIntakeRequest"
      responses:
        "200":
          description: Draft updated from intake.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftMutateResponse"
  /v1/workflow-drafts/{draftId}/synthesize:
    post:
      tags: [Workflow Drafts]
      summary: Synthesize a draft into a publish-ready spec candidate.
      description: Authoring helper only. Not production ingress.
      parameters:
        - $ref: "#/components/parameters/draftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftFocusRequest"
      responses:
        "200":
          description: Draft synthesized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftMutateResponse"
  /v1/workflow-drafts/{draftId}/validate:
    post:
      tags: [Workflow Drafts]
      summary: Validate a draft against pure-v1 publishing rules.
      parameters:
        - $ref: "#/components/parameters/draftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftFocusRequest"
      responses:
        "200":
          description: Draft validated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftMutateResponse"
  /v1/workflow-drafts/{draftId}/publish:
    post:
      tags: [Workflow Drafts]
      summary: Publish a draft into an authoritative workflow template version.
      parameters:
        - $ref: "#/components/parameters/draftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftPublishRequest"
      responses:
        "200":
          description: Draft published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftPublishResponse"
  /v1/workflow-drafts/{draftId}/spec:
    patch:
      tags: [Workflow Drafts]
      summary: Patch the current authoring spec for a draft.
      parameters:
        - $ref: "#/components/parameters/draftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDraftSpecPatchRequest"
      responses:
        "200":
          description: Draft spec patched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraftSpecPatchResponse"
  /v1/agents:
    get:
      tags: [Agents]
      summary: List agent definitions.
      responses:
        "200":
          description: Agent list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDefinitionListResponse"
    post:
      tags: [Agents]
      summary: Create an agent definition.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDefinitionCreateRequest"
      responses:
        "200":
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDefinitionResponse"
  /v1/agents/{agentId}:
    get:
      tags: [Agents]
      summary: Get an agent definition.
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDefinitionResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentId}/runs:
    post:
      tags: [Runs, Agents]
      summary: Start a production run from an authoritative agent.
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunStartRequest"
      responses:
        "200":
          description: Run started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowCommandResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/agents/{agentId}/trigger-bindings:
    get:
      tags: [Trigger Bindings, Agents]
      summary: List trigger bindings owned by an agent.
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Trigger binding list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerBindingListResponse"
    post:
      tags: [Trigger Bindings, Agents]
      summary: Create a trigger binding for an agent.
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerBindingCreateRequest"
      responses:
        "200":
          description: Trigger binding created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerBindingResponse"
  /v1/trigger-bindings/{triggerBindingId}/enable:
    post:
      tags: [Trigger Bindings]
      summary: Enable a trigger binding.
      parameters:
        - $ref: "#/components/parameters/triggerBindingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerBindingLifecycleRequest"
      responses:
        "200":
          description: Trigger binding enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerBindingResponse"
  /v1/trigger-bindings/{triggerBindingId}/disable:
    post:
      tags: [Trigger Bindings]
      summary: Disable a trigger binding.
      parameters:
        - $ref: "#/components/parameters/triggerBindingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerBindingLifecycleRequest"
      responses:
        "200":
          description: Trigger binding disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerBindingResponse"
  /v1/runs:
    get:
      tags: [Runs]
      summary: List recent workflow runs.
      responses:
        "200":
          description: Run list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunListResponse"
    post:
      tags: [Runs]
      summary: Start a direct studio/debug run from a published workflow version.
      description: Non-production entry path. Use only for studio/debug flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowVersionRunStartRequest"
      responses:
        "200":
          description: Run started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowCommandResponse"
  /v1/runs/{runId}:
    get:
      tags: [Runs]
      summary: Get a workflow run snapshot.
      parameters:
        - $ref: "#/components/parameters/runId"
      responses:
        "200":
          description: Run detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunQueryResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/runs/{runId}/cancel:
    post:
      tags: [Runs]
      summary: Cancel an in-flight workflow run.
      parameters:
        - $ref: "#/components/parameters/runId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowRunCancelRequest"
      responses:
        "200":
          description: Run cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowCommandResponse"
...
  /v1/approvals/{approvalRequestId}:
    get:
      tags: [Approvals]
      summary: Get approval detail.
      parameters:
        - $ref: "#/components/parameters/approvalRequestId"
      responses:
        "200":
          description: Approval detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowApprovalDetailResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/approvals/{approvalRequestId}/decision:
    post:
      tags: [Approvals]
      summary: Decide an approval request.
      parameters:
        - $ref: "#/components/parameters/approvalRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowApprovalDecisionRequest"
      responses:
        "200":
          description: Approval decision applied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowApprovalDecisionResponse"
  /v1/interactions/{interactionRequestId}/responses:
    post:
      tags: [Interactions]
      summary: Submit an interaction response.
      description: The authoritative interaction resume path. Generic run resume is removed from the mainline API.
      parameters:
        - $ref: "#/components/parameters/interactionRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInteractionResponseRequest"
      responses:
        "200":
          description: Interaction response accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInteractionResponseResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/artifacts/{artifactId}:
    get:
      tags: [Artifacts]
      summary: Get artifact detail.
      parameters:
        - $ref: "#/components/parameters/artifactId"
      responses:
        "200":
          description: Artifact detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowArtifactDetailResponse"
        "404":
          $ref: "#/components/responses/Error"
  /v1/bridge-registrations:
    get:
      tags: [Bridge Registrations]
      summary: List external runtime bridge registrations.
      responses:
        "200":
          description: Bridge list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeRegistrationListResponse"
    post:
      tags: [Bridge Registrations]
      summary: Create an external runtime bridge registration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BridgeRegistrationCreateRequest"
      responses:
        "200":
          description: Bridge created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeRegistrationResponse"
  /v1/bridge-registrations/{bridgeId}:
    get:
      tags: [Bridge Registrations]
      summary: Get an external runtime bridge registration.
      parameters:
        - $ref: "#/components/parameters/bridgeId"
      responses:
        "200":
          description: Bridge detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeRegistrationResponse"
  /v1/connector-definitions:
    get:
      tags: [Connectors]
      summary: List connector definitions.
      responses:
        "200":
          description: Connector definition list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorDefinitionListResponse"
    post:
      tags: [Connectors]
      summary: Create a connector definition.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorDefinitionCreateRequest"
      responses:
        "200":
          description: Connector definition created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorDefinitionResponse"
  /v1/connector-bindings:
    get:
      tags: [Connectors]
      summary: List connector bindings.
      responses:
        "200":
          description: Connector binding list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorBindingListResponse"
    post:
      tags: [Connectors]
      summary: Create a connector binding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorBindingCreateRequest"
      responses:
        "200":
          description: Connector binding created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorBindingResponse"
  /v1/agents/{agentId}/action-bindings:
    get:
      tags: [Connectors, Agents]
      summary: List action bindings for an agent.
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Action binding list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionBindingListResponse"
    post:
      tags: [Connectors, Agents]
      summary: Create an action binding for an agent.
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionBindingCreateRequest"
      responses:
        "200":
          description: Action binding created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionBindingResponse"
  /v1/event-subscriptions:
    get:
      tags: [Connectors]
      summary: List event subscriptions.
      responses:
        "200":
          description: Event subscription list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSubscriptionListResponse"
    post:
      tags: [Connectors]
      summary: Create an event subscription.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventSubscriptionCreateRequest"
      responses:
        "200":
          description: Event subscription created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventSubscriptionResponse"
  /v1/policy-bindings:
    get:
      tags: [Policy]
      summary: List policy bindings.
      responses:
        "200":
          description: Policy binding list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyBindingListResponse"
    post:
      tags: [Policy]
      summary: Create a policy binding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyBindingCreateRequest"
      responses:
        "200":
          description: Policy binding created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyBindingResponse"
  /v1/secret-refs:
    get:
      tags: [Policy]
      summary: List secret refs.
      responses:
        "200":
          description: Secret ref list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretRefListResponse"
    post:
      tags: [Policy]
      summary: Create a secret ref.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretRefCreateRequest"
      responses:
        "200":
          description: Secret ref created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretRefResponse"
...
  /v1/governance-change-requests:
    get:
      tags: [Governance Change Proposals]
      summary: List governance-change proposals.
      description: Auxiliary proposal storage only. Not authoritative runtime approval state.
      responses:
        "200":
          description: Governance-change request list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceChangeRequestListResponse"
    post:
      tags: [Governance Change Proposals]
      summary: Create a governance-change proposal.
      description: Auxiliary proposal storage only. Not authoritative runtime approval state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernanceChangeRequestCreateRequest"
      responses:
        "200":
          description: Governance-change proposal created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceChangeRequestResponse"
  /v1/governance-change-requests/{requestId}:
    get:
      tags: [Governance Change Proposals]
      summary: Get a governance-change proposal.
      description: Auxiliary proposal storage only. Not authoritative runtime approval state.
      parameters:
        - $ref: "#/components/parameters/requestId"
      responses:
        "200":
          description: Governance-change proposal detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceChangeRequestDetailResponse"
...
components:
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
    workflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string }
    draftId:
      name: draftId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    WorkflowDetailResponse:
      type: object
      properties:
        workflow:
          $ref: "#/components/schemas/WorkflowTemplateRecord"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowTemplateVersionRecord"
      required: [workflow, versions]
...
    WorkflowDraftListResponse:
      type: object
      properties:
        schemaVersion: { type: string, enum: [v1] }
        drafts:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowDraftRecord"
        sessionLinks:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowDraftSessionLinkRecord"
      required: [schemaVersion, drafts, sessionLinks]
...
    WorkflowDraftCreateRequest:
      type: object
      properties:
        schemaVersion: { type: string, enum: [v1] }
        sessionId: { type: string }
        userId: { type: string }
        workflowKey: { type: string }
        name: { type: string }
        basedOnTemplateVersionId: { type: string }
        source: { type: string }
        initialText: { type: string }
      required: [schemaVersion, sessionId, userId]
...
    WorkflowDraftSpecPatchResponse:
      allOf:
        - $ref: "#/components/schemas/WorkflowDraftMutateResponse"
...